Following [a suggestion](https://bsky.app/profile/snek.dev/post/3lb5tbnoy3k2d) I am introducing a new dependency on [Binaryen](https://github.com/WebAssembly/binaryen)'s [npm package](https://www.npmjs.com/package/binaryen) and adding a [`wasm-opt`](https://github.com/WebAssembly/binaryen?tab=readme-ov-file#wasm-opt) step to the build process. Currently I'm using `-O4` because this is an extremely small piece of code so I see no reason not to apply the maximum possible amount of optimisation and inlining and suchlike.

This massively improves the WASM's performance. With *n* = 23:

## Before

```
HK/JS: 6.414s
HK/WASM: 15.227s
```

## After

```
HK/JS: 6.090s
HK/WASM: 7.824s
```

So, around 100% faster! (Vaguely deterministic performance tests might be an idea at some point though.) But still not at the same level as the JavaScript! Maybe TurboFan is just beyond our capabilities?